+++ gazebo/gui/model/MEUserCmdManager.cc
@@ -47,6 +47,12 @@ namespace gazebo
       /// \brief If the command is related to a joint, this is its unique Id.
       /// It's different from the scopedName and we need both.
       public: std::string jointId;
+
+      /// \brief Pose before the command (to be used by undo).
+      public: ignition::math::Pose3d poseBefore;
+
+      /// \brief Pose after the command (to be used by redo).
+      public: ignition::math::Pose3d poseAfter;
     };
 
     /// \internal
@@ -130,6 +136,20 @@ void MEUserCmd::Undo()
 
     model::Events::requestJointInsertion(this->dataPtr->sdf, topModelName);
   }
+  // Moving a link
+  else if (this->dataPtr->type == MEUserCmd::MOVING_LINK &&
+      !this->dataPtr->scopedName.empty())
+  {
+    model::Events::requestLinkMove(this->dataPtr->scopedName,
+        this->dataPtr->poseBefore);
+  }
+  // Moving a nested model
+  else if (this->dataPtr->type == MEUserCmd::MOVING_NESTED_MODEL &&
+      !this->dataPtr->scopedName.empty())
+  {
+    model::Events::requestNestedModelMove(this->dataPtr->scopedName,
+        this->dataPtr->poseBefore);
+  }
 }
 
 /////////////////////////////////////////////////
@@ -176,6 +196,20 @@ void MEUserCmd::Redo()
   {
     model::Events::requestJointRemoval(this->dataPtr->jointId);
   }
+  // Moving a link
+  else if (this->dataPtr->type == MEUserCmd::MOVING_LINK &&
+      !this->dataPtr->scopedName.empty())
+  {
+    model::Events::requestLinkMove(this->dataPtr->scopedName,
+        this->dataPtr->poseAfter);
+  }
+  // Moving a nested model
+  else if (this->dataPtr->type == MEUserCmd::MOVING_NESTED_MODEL &&
+      !this->dataPtr->scopedName.empty())
+  {
+    model::Events::requestNestedModelMove(this->dataPtr->scopedName,
+        this->dataPtr->poseAfter);
+  }
 }
 
 /////////////////////////////////////////////////
@@ -206,6 +240,14 @@ void MEUserCmd::SetScopedName(const std::string &_name)
 void MEUserCmd::SetJointId(const std::string &_id)
 {
   this->dataPtr->jointId = _id;
+}
+
+/////////////////////////////////////////////////
+void MEUserCmd::SetPoseChange(const ignition::math::Pose3d &_before,
+    const ignition::math::Pose3d &_after)
+{
+  this->dataPtr->poseBefore = _before;
+  this->dataPtr->poseAfter = _after;
 }
 
 /////////////////////////////////////////////////