+++ test/integration/model_editor_undo.hh
@@ -41,6 +41,12 @@ class ModelEditorUndoTest : public QTestFixture
   /// \brief Test undo/redo joint insertion via creation dialog.
   private slots: void JointInsertionByDialog();
 
+  /// \brief Test undo/redo link translation.
+  private slots: void LinkTranslation();
+
+  /// \brief Test undo/redo nested model align.
+  private slots: void NestedModelAlign();
+
   /// \brief Helper callback to trigger the delete action on the context
   /// menu after the menu, which is modal, has been opened.
   private slots: void TriggerDelete();